# be/swagger.yaml
...
openapi: 3.0.3
info:
  title: 쉐어스토리 Backend API
  version: 1.0.0
  description: |-
    be/src/app.js에 실제 연결된 라우트와 컨트롤러/서비스/모델 기준 문서입니다.

    인증: POST /auth/login의 token을 Swagger Authorize에 입력하세요. Bearer 접두사는 자동으로 붙습니다. 역할은 계정의 고정 속성이 아니라 모임별 leader_id/Apply 관계입니다.

    응답: 대부분 { success, document }를 사용하지만 아이디 중복 확인은 { message, count }입니다. 공통 errorHandler는 success 없이 { message }만 반환합니다. 날짜/시간 문자열과 BIGINT 문자열을 구분하세요.

    현재 구현 범위: 세션 배치 router는 app.js에 등록되어 있지 않아 호출 가능한 paths에 포함하지 않았습니다. 파일 업로드 middleware도 라우트에 연결되어 있지 않습니다. 로그북 승인, 리뷰 수정/삭제, 모임 종료 처리 API는 없습니다. complete-meetups.job.js는 비어 있습니다.

    실행 참고: 현재 meetup.service.js가 참조하는 common/services/mailer.service.js가 누락되어 앱 시작을 막습니다. swagger-ui-express 의존성은 있지만 Swagger UI 경로는 app.js에 등록되어 있지 않습니다. 이 파일은 Swagger Editor/UI에서 별도로 불러올 수 있습니다. 회원정보 수정 API의 기존 오류는 해당 작업 설명을 확인하세요.

    캡틴 화면 조회 순서: GET /member/meetups/captain → 모임 선택 → GET /member/meetups/{meetup_id}/crews 및 /sessions → 세션 선택 → GET /logbook/meetups/{meetup_id}/sessions/{session_id}. 마지막 응답의 크루를 선택하여 logbook을 표시하며 null이면 미제출 상태를 표시합니다.
servers:
  - url: http://localhost:{port}
    description: 로컬 서버 (PORT 환경변수에 맞게 설정)
    variables:
      port:
        default: '3000'
...
paths:
...
  '/auth/updateMyInfo':
    patch:
      tags:
        - Auth
      operationId: updateMyInfo
      summary: 회원정보 수정 (현재 구현 오류)
      description:
        현재 인증이 없고 body.user_id 대상 정보를 먼저 수정합니다. 이후 User.update 반환 배열에서
        password를 읽어 bcrypt.compare를 호출하므로 정상 성공을 보장하지 못하며, 성공 분기에도 미정의 token을 참조합니다.
        500이어도 DB 변경이 반영될 수 있습니다. 정상 200 응답 계약은 문서화하지 않습니다.
      security: []
      responses:
        '400':
          '$ref': '#/components/responses/Error400'
        '409':
          '$ref': '#/components/responses/Error409'
        '500':
          '$ref': '#/components/responses/Error500'
        '504':
          '$ref': '#/components/responses/Error504'
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/UpdateMyInfoRequest'
  '/meetup':
    post:
      tags:
        - Meetup
      operationId: createMeetup
      summary: 모임 및 4개 세션 생성
      description:
        로그인 사용자가 캡틴입니다. 모임은 RECRUITING, 세션은 SCHEDULED로 생성됩니다. 예시 deadline은
        실행 시 현재 이후 날짜로 변경하세요.
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  document:
                    '$ref': '#/components/schemas/SavedMeetupDocument'
                  message:
                    type: string
                required:
                  - success
                  - document
                  - message
        '400':
          '$ref': '#/components/responses/Error400'
        '404':
          '$ref': '#/components/responses/Error404'
        '500':
          '$ref': '#/components/responses/Error500'
        '401':
          '$ref': '#/components/responses/Error401'
        '403':
          '$ref': '#/components/responses/Error403'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateMeetupRequest'
            example:
              title: 함께 읽는 독서 모임
              description: 네 번의 독서 토론을 진행합니다.
              book_title: 함께 읽을 책
              price: 10000
              min_capacity: 4
              max_capacity: 8
              deadline: '2026-11-01T00:00:00Z'
              sessions:
                - session_number: 1
                  topic: 1회차 독서 토론
                  sch_date: '2026-11-07'
                  sch_day: 토요일
                  sch_time: '19:00'
                  sch_st_time: '19:00'
                  sch_ed_time: '21:00'
                - session_number: 2
                  topic: 2회차 독서 토론
                  sch_date: '2026-11-14'
                  sch_day: 토요일
                  sch_time: '19:00'
                  sch_st_time: '19:00'
                  sch_ed_time: '21:00'
                - session_number: 3
                  topic: 3회차 독서 토론
                  sch_date: '2026-11-21'
                  sch_day: 토요일
                  sch_time: '19:00'
                  sch_st_time: '19:00'
                  sch_ed_time: '21:00'
                - session_number: 4
                  topic: 4회차 독서 토론
                  sch_date: '2026-11-28'
                  sch_day: 토요일
                  sch_time: '19:00'
                  sch_st_time: '19:00'
                  sch_ed_time: '21:00'
    get:
      tags:
        - Meetup
      operationId: listMeetups
      summary: 공개 모임 목록
      description: 취소되지 않은 세션의 회차 순서로 시작/종료 일정을 계산합니다.
      security: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  document:
                    type: object
                    properties:
                      page:
                        type: integer
                        minimum: 1
                      limit:
                        type: integer
                        minimum: 1
                        maximum: 100
                      total:
                        type: integer
                        minimum: 0
                      nextPage:
                        type: integer
                        nullable: true
                      items:
                        type: array
                        items:
                          '$ref': '#/components/schemas/MeetupListItem'
                    required:
                      - page
                      - limit
                      - total
                      - nextPage
                      - items
                required:
                  - success
                  - document
        '500':
          '$ref': '#/components/responses/Error500'
      parameters:
        - name: page
          in: query
          description: 잘못된 숫자는 1로 보정합니다.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 잘못된 숫자는 10, 100 초과는 100으로 보정합니다.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: keyword
          in: query
          description: 모임 제목 부분 검색 (대소문자 무시)
          schema:
            type: string
        - name: status
          in: query
          description: ''
          schema:
            '$ref': '#/components/schemas/MeetupStatus'
  '/meetup/{meetup_id}':
    get:
      tags:
        - Meetup
      operationId: getMeetupDetail
      summary: 공개 모임 상세
      description: 세션 일정과 신청자 독서 성향 통계를 반환합니다. Zoom 접속 정보는 포함하지 않습니다.
      security: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  document:
                    '$ref': '#/components/schemas/MeetupDetail'
                required:
                  - success
                  - document
        '400':
          '$ref': '#/components/responses/Error400'
        '404':
          '$ref': '#/components/responses/Error404'
        '500':
          '$ref': '#/components/responses/Error500'
      parameters:
        - name: meetup_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[1-9][0-9]*$'
            example: '1'
          description:
            양의 정수 식별자. member/logbook/review 및 신청 API는 최대 9223372036854775807까지
            문자열로 처리합니다. 기존 모임 상세/수정은 Number로 변환합니다.
    patch:
      tags:
        - Meetup
      operationId: updateMeetup
      summary: 캡틴의 모임 및 세션 수정
      description:
        캡틴만 수정 가능합니다. 세션 수정 시 sch_time이 sch_st_time과 sch_ed_time에 모두 저장되는
        현재 구현에 유의하세요.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  document:
                    '$ref': '#/components/schemas/SavedMeetupDocument'
                  message:
                    type: string
                required:
                  - success
                  - document
                  - message
        '400':
          '$ref': '#/components/responses/Error400'
        '404':
          '$ref': '#/components/responses/Error404'
        '500':
          '$ref': '#/components/responses/Error500'
        '401':
          '$ref': '#/components/responses/Error401'
        '403':
          '$ref': '#/components/responses/Error403'
      parameters:
        - name: meetup_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[1-9][0-9]*$'
            example: '1'
          description:
            양의 정수 식별자. member/logbook/review 및 신청 API는 최대 9223372036854775807까지
            문자열로 처리합니다. 기존 모임 상세/수정은 Number로 변환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/UpdateMeetupRequest'
            example:
              meetup:
                title: 변경된 독서 모임
  '/meetup/{meetup_id}/apply':
    post:
      tags:
        - Meetup
      operationId: applyMeetup
      summary: 모임 신청
      description:
        member/meetup.controller가 처리합니다. 요청 본문 없이 로그인 사용자로 신청하며 Apply 전체를
        반환합니다(payment_status 없음). 모집 중이고 마감 전이며 정원이 남아 있어야 합니다. 캡틴의 자기 모임 신청은 400,
        중복/마감/정원 초과는 409입니다. 별도 승인·결제 없이 ING 상태로 참여합니다.
      security:
        - bearerAuth: []
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  document:
                    '$ref': '#/components/schemas/Apply'
                required:
                  - success
                  - document
        '400':
          '$ref': '#/components/responses/Error400'
        '404':
          '$ref': '#/components/responses/Error404'
        '409':
          '$ref': '#/components/responses/Error409'
        '500':
          '$ref': '#/components/responses/Error500'
        '504':
          '$ref': '#/components/responses/Error504'
        '401':
          '$ref': '#/components/responses/Error401'
        '403':
          '$ref': '#/components/responses/Error403'
      parameters:
        - name: meetup_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[1-9][0-9]*$'
            example: '1'
          description:
            양의 정수 식별자. member/logbook/review 및 신청 API는 최대 9223372036854775807까지
            문자열로 처리합니다. 기존 모임 상세/수정은 Number로 변환합니다.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: POST /auth/login에서 반환된 token을 입력합니다.
  responses:
    Error400:
      description: 입력값/DB 검증 실패
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
    Error401:
      description: Bearer 인증 헤더 누락 또는 로그인 사용자 식별자 없음
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
    Error403:
      description: 유효하지 않은 토큰 또는 권한 없음
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
    Error404:
      description: 대상 리소스 없음
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
    Error409:
      description: 중복 데이터 또는 현재 상태에서 처리 불가
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
    Error500:
      description: 서버 또는 데이터베이스 오류
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
    Error504:
      description: 데이터베이스 쿼리 시간 초과
      content:
        application/json:
          schema:
            '$ref': '#/components/schemas/Error'
  schemas:
    BigId:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '1'
      description: PostgreSQL BIGINT 식별자는 JSON 응답에서 문자열로 반환됩니다.
    MeetupStatus:
      type: string
      enum:
        - RECRUITING
        - CLOSED
        - IN_PROGRESS
        - COMPLETED
    SessionStatus:
      type: string
      enum:
        - SCHEDULED
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED
    ApplyStatus:
      type: string
      enum:
        - ING
        - COMPLETED
      description: Apply 모델의 실제 상태. Enum.js에 남아 있는 PENDING/APPROVED 등은 사용하지 않습니다.
    Genre:
      type: string
      enum:
        - NOVEL
        - ECONOMY_BUSINESS
        - SELF_DEVELOPMENT
        - IT
        - ESSAY
        - TRAVEL_LIFESTYLE
        - PARENT_CHILD
        - HUMANITIES_PHILOSOPHY
        - SOCIETY
        - SCIENCE
        - HISTORY
        - ETC
      nullable: true
    ReadingAmount:
      type: string
      enum:
        - BOOKS_1_2
        - BOOKS_3_4
        - BOOKS_5_6
        - BOOKS_7_PLUS
...
    PublicSession:
      type: object
      properties:
        session_id:
          '$ref': '#/components/schemas/BigId'
        session_number:
          type: integer
          minimum: 1
        topic:
          type: string
          maxLength: 200
        sch_date:
          type: string
          example: '2026-11-07'
        sch_day:
          type: string
          maxLength: 10
          example: 토요일
        sch_time:
          type: string
          example: '19:00'
        sch_st_time:
          type: string
          example: '19:00'
        sch_ed_time:
          type: string
          example: '21:00'
        status:
          '$ref': '#/components/schemas/SessionStatus'
    SavedSession:
      type: object
      properties:
        session_id:
          '$ref': '#/components/schemas/BigId'
        meetup_id:
          '$ref': '#/components/schemas/BigId'
        session_number:
          type: integer
          minimum: 1
        topic:
          type: string
          maxLength: 200
        sch_date:
          type: string
          example: '2026-11-07'
        sch_day:
          type: string
          maxLength: 10
          example: 토요일
        sch_time:
          type: string
          example: '19:00'
        zoom_url:
          type: string
          nullable: true
          maxLength: 512
        zoom_password:
          type: string
          nullable: true
          maxLength: 50
        status:
          '$ref': '#/components/schemas/SessionStatus'
    SavedMeetupDocument:
      type: object
      properties:
        meetup:
          type: object
          properties:
            meetup_id:
              '$ref': '#/components/schemas/BigId'
            leader_id:
              type: string
            title:
              type: string
              maxLength: 100
            description:
              type: string
            book_title:
              type: string
              maxLength: 150
            book_image_url:
              type: string
              maxLength: 512
              nullable: true
            price:
              type: integer
              minimum: 0
            min_capacity:
              type: integer
              minimum: 1
            max_capacity:
              type: integer
            deadline:
              type: string
              format: date-time
            status:
              '$ref': '#/components/schemas/MeetupStatus'
            created_at:
              type: string
              format: date-time
        sessions:
          type: array
          items:
            '$ref': '#/components/schemas/SavedSession'
      required:
        - meetup
        - sessions
    MeetupListItem:
      type: object
      properties:
        meetup_id:
          '$ref': '#/components/schemas/BigId'
        title:
          type: string
          maxLength: 100
        book_title:
          type: string
          maxLength: 150
        book_image_url:
          type: string
          maxLength: 512
          nullable: true
        description:
          type: string
        min_capacity:
          type: integer
          minimum: 1
        max_capacity:
          type: integer
        deadline:
          type: string
          format: date-time
        status:
          '$ref': '#/components/schemas/MeetupStatus'
        leader_name:
          type: string
          nullable: true
        sch_st_date:
          type: string
          nullable: true
        sch_ed_date:
          type: string
          nullable: true
        sch_day:
          type: string
          nullable: true
        sch_time:
          type: string
          nullable: true
    MeetupDetail:
      type: object
      properties:
        meetup:
          type: object
          properties:
            meetup_id:
              '$ref': '#/components/schemas/BigId'
            leader_id:
              type: string
            title:
              type: string
              maxLength: 100
            description:
              type: string
            book_title:
              type: string
              maxLength: 150
            book_image_url:
              type: string
              maxLength: 512
              nullable: true
            price:
              type: integer
              minimum: 0
            min_capacity:
              type: integer
              minimum: 1
            max_capacity:
              type: integer
            deadline:
              type: string
              format: date-time
            status:
              '$ref': '#/components/schemas/MeetupStatus'
            leader_name:
              type: string
              nullable: true
        sessions:
          type: array
          items:
            '$ref': '#/components/schemas/PublicSession'
        apply_count:
          type: integer
          minimum: 0
        apply_user_stats:
          type: object
          properties:
            genre_1:
              type: object
              additionalProperties:
                type: integer
                minimum: 0
            genre_2:
              type: object
              additionalProperties:
                type: integer
                minimum: 0
            monthly_reading_volume:
              type: object
              additionalProperties:
                type: integer
                minimum: 0
            age_group:
              type: object
              additionalProperties:
                type: integer
                minimum: 0
            gender:
              type: object
              additionalProperties:
                type: integer
                minimum: 0
    Apply:
      type: object
      properties:
        apply_id:
          '$ref': '#/components/schemas/BigId'
        meetup_id:
          '$ref': '#/components/schemas/BigId'
        user_id:
          type: string
        status:
          '$ref': '#/components/schemas/ApplyStatus'
        created_user_id:
          type: string
        updated_user_id:
          type: string
        deleted_user_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    UpdateMyInfoRequest:
      type: object
      properties:
        user_id:
          type: string
          maxLength: 100
        email:
          type: string
          maxLength: 100
        name:
          type: string
          maxLength: 50
        gender:
          type: string
          enum:
            - M
            - F
        age_group:
          type: string
          maxLength: 100
        monthly_reading_volume:
          '$ref': '#/components/schemas/ReadingAmount'
        genre_1:
          '$ref': '#/components/schemas/Genre'
        genre_2:
          '$ref': '#/components/schemas/Genre'
        password:
          type: string
          description: 현재 구현의 비밀번호 비교 단계에 사용되나 처리 오류가 있습니다.
      description: 현재 인증 미들웨어가 없습니다. 업데이트가 먼저 수행된 후 비밀번호 비교 단계에서 오류가 발생할 수 있습니다.
      required:
        - user_id
    CreateSessionRequest:
      type: object
      properties:
        session_number:
          type: integer
          minimum: 1
        topic:
          type: string
          maxLength: 200
        sch_date:
          type: string
          example: '2026-11-07'
        sch_day:
          type: string
          maxLength: 10
          example: 토요일
        sch_time:
          type: string
          example: '19:00'
        sch_st_time:
          type: string
          example: '19:00'
        sch_ed_time:
          type: string
          example: '21:00'
        zoom_url:
          type: string
          nullable: true
          maxLength: 512
        zoom_password:
          type: string
          nullable: true
          maxLength: 50
      required:
        - session_number
        - topic
        - sch_date
        - sch_day
        - sch_time
        - sch_st_time
        - sch_ed_time
    CreateMeetupRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 1
        description:
          type: string
          minLength: 1
        book_title:
          type: string
          maxLength: 150
          minLength: 1
        book_image_url:
          type: string
          maxLength: 512
          nullable: true
        price:
          type: integer
          minimum: 0
        min_capacity:
          type: integer
          minimum: 4
          maximum: 8
        max_capacity:
          type: integer
          minimum: 4
          maximum: 8
        deadline:
          type: string
          format: date-time
        sessions:
          type: array
          items:
            '$ref': '#/components/schemas/CreateSessionRequest'
          minItems: 4
          maxItems: 4
      description: min_capacity <= max_capacity, deadline은 현재 이후, session_number는
        중복 불가. leader_id는 JWT에서 결정합니다.
      required:
        - title
        - description
        - book_title
        - price
        - min_capacity
        - max_capacity
        - deadline
        - sessions
    UpdateSessionRequest:
      type: object
      properties:
        session_number:
          type: integer
          minimum: 1
        topic:
          type: string
          maxLength: 200
        sch_date:
          type: string
          example: '2026-11-07'
        sch_day:
          type: string
          maxLength: 10
          example: 토요일
        sch_time:
          type: string
          example: '19:00'
        sch_st_time:
          type: string
          example: '19:00'
        sch_ed_time:
          type: string
          example: '21:00'
        zoom_url:
          type: string
          nullable: true
          maxLength: 512
        zoom_password:
          type: string
          nullable: true
          maxLength: 50
        session_id:
          type: integer
          minimum: 1
        meetup_id:
          type: integer
          minimum: 1
      description:
        세션 수정에도 일정 필드 전체를 검증합니다. sch_st_time/sch_ed_time은 입력을 직접 저장하지 않고
        sch_time 값으로 둘 다 덮어씁니다. session_id는 숫자로 전달합니다.
      required:
        - session_id
        - session_number
        - topic
        - sch_date
        - sch_day
        - sch_time
        - sch_st_time
        - sch_ed_time
    UpdateMeetupRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
        book_title:
          type: string
          maxLength: 150
        book_image_url:
          type: string
          maxLength: 512
          nullable: true
        price:
          type: integer
          minimum: 0
        min_capacity:
          type: integer
          minimum: 1
        max_capacity:
          type: integer
        deadline:
          type: string
          format: date-time
        meetup_id:
          type: integer
          minimum: 1
        meetup:
          type: object
          properties:
            title:
              type: string
              maxLength: 100
            description:
              type: string
            book_title:
              type: string
              maxLength: 150
            book_image_url:
              type: string
              maxLength: 512
              nullable: true
            price:
              type: integer
              minimum: 0
            min_capacity:
              type: integer
              minimum: 1
            max_capacity:
              type: integer
            deadline:
              type: string
              format: date-time
          description: 있으면 최상위 모임 필드 대신 이 객체를 사용합니다.
        sessions:
          type: array
          items:
            '$ref': '#/components/schemas/UpdateSessionRequest'
      description:
        모임 필드는 부분 수정 가능. status/leader_id 변경은 지원하지 않습니다. sessions는 반드시
        최상위에 전달합니다.
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - false
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      description: 직접 처리하는 오류에는 success=false가 포함됩니다. 공통 errorHandler는 message만 반환합니다.
      required:
        - message
...